+++ src/main/kotlin/models/CounterTypes.kt
@@ -1,0 +1,7 @@
+package net.razvan.models
+
+object CounterTypes {
+    val BRANCH = "BRANCH"
+    val LINE = "LINE"
+    val COMPLEXITY = "COMPLEXITY"
+}