+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This list is not intended to be all-encompassing - it will document major and breaking changes with their rationale when appropriate:
 
+### v2.0.0
+- Replace XML library with Jackson for better performance(?) and maintainability
+- Fixed a number of issues with the XML generated that now is DTD compliant
+- Update to gradle version catalogs
+
+
 ### v1.3.2
 - fix #11: report.dtd not found